requirements.txt

numpy
h5py
tqdm
psutil
pyyaml
matplotlib
ruff